feat: constructor support for the autocrafting upgrade
When there is already a block in front of the constructor
(and thus nothing is required to happen)
RS will still create an extra set even when
not necessary because it first tries to extract
before checking if anything needs to happen.
This is difficult to work around because
placing a block does not support
simulations.