fix: not being able to insert resources in a grid when clicking on autocraftable resource
We need to give priority to insertions before autocrafting.
However, extractions need to be checked first
so that we can still fill buckets by clicking
an empty bucket on a fluid for example!
We cannot simply give priority to all insertions
when holding an item as otherwise that case would fail.

Fixes #806