fix: item count rendering behind the item in the interface slots
Because the slot has type CONTAINER, renderSlot
was interfering and rendering the
counts correctly for items when count>1.

However, when count==1,
renderSlot does not kick in and we can see
the effects of our own renderAmount function,
and that was rendering at a wrong Z-level
for some reason.

The Z-level was changed in 838a5cad but it's unclear why.

Fixes #839