fix: constructor starting autocrafting when block is already placed
Block placement has no simulate option,
so we always try to place something,
however, when we always try to place
something, we always have to check if
the resource is available.
If we always check if the resource is available before we try to place,
we might trigger autocrafting.
So, we stop early if there is already
a block there.
This might break cases like placing
fences in water.
This avoids always keeping an extra
set in stock when unneeded.